+++ sys/src/ape/lib/ap/stdio/ftell.c
@@ -2,8 +2,8 @@
  * pANS stdio -- ftell
  */
 #include "iolib.h"
-long ftell(FILE *f){
-	long seekp=lseek(f->fd, 0L, 1);
+long long ftell(FILE *f){
+	long long seekp=lseek(f->fd, 0L, 1);
 	if(seekp<0) return -1;		/* enter error state? */
 	switch(f->state){
 	default:
